add session database

+++ app/src/main/java/com/talkie/wtalkie/ui/ContactsFragment.java
@@ -33,7 +33,7 @@ public class ContactsFragment extends Fragment implements AdapterView.OnItemClic
     private boolean mChecked = false;
     private MyBaseAdapter mAdapter;
 
-    private int[] mUserIds;
+    private long[] mUserIds;
 
 
 /* ********************************************************************************************** */
@@ -95,7 +95,7 @@ public class ContactsFragment extends Fragment implements AdapterView.OnItemClic
         // state machine
         if (mState == SELECT_STATE_IDLE){
             Intent intent = new Intent(this.getActivity(), ChatActivity.class);
-            mUserIds = new int[1];
+            mUserIds = new long[1];
             mUserIds[0] = position + 1;
             intent.putExtra("UserIds", mUserIds);
             startActivity(intent);
@@ -119,11 +119,10 @@ public class ContactsFragment extends Fragment implements AdapterView.OnItemClic
 
             Intent intent = new Intent(this.getActivity(), ChatActivity.class);
             int count = Users.getUsersCount();
-            mUserIds = new int[count];
+            mUserIds = new long[count];
             for (int i = 0; i < count; i++){
                 mUserIds[i] = i + 1;
             }
-
             intent.putExtra("UserIds", mUserIds);
             startActivity(intent);
